Moved more bench files
