+++ sei-db/state_db/bench/bench_ss_test.go
@@ -20,26 +20,7 @@ func BenchmarkSSCompositeWrite(b *testing.B) {
 
 	for _, scenario := range scenarios {
 		b.Run(scenario.Name, func(b *testing.B) {
-			runBenchmark(b, scenario, false)
-		})
-	}
-}
-
-func BenchmarkSSHistoricalOffloadWrite(b *testing.B) {
-	const totalKeys int64 = 10_000
-
-	scenarios := []TestScenario{
-		{
-			Name:      "100_keys_per_block",
-			TotalKeys: totalKeys,
-			NumBlocks: totalKeys / 100,
-			Backend:   wrappers.SSHistoricalOffload,
-		},
-	}
-
-	for _, scenario := range scenarios {
-		b.Run(scenario.Name, func(b *testing.B) {
-			runBenchmark(b, scenario, false)
+			runBenchmark(b, scenario)
 		})
 	}
 }
@@ -58,7 +39,7 @@ func BenchmarkCombinedCompositeDualSSCompositeWrite(b *testing.B) {
 
 	for _, scenario := range scenarios {
 		b.Run(scenario.Name, func(b *testing.B) {
-			runBenchmark(b, scenario, false)
+			runBenchmark(b, scenario)
 		})
 	}
 }
