+++ sei-db/tools/cmd/seidb/benchmark/write.go
@@ -4,10 +4,16 @@ import (
 	"fmt"
 	"io/fs"
 	"os"
+	"runtime"
+	"sort"
+	"sync"
+	"time"
 
 	"github.com/sei-protocol/sei-chain/sei-db/config"
+	"github.com/sei-protocol/sei-chain/sei-db/db_engine/types"
+	"github.com/sei-protocol/sei-chain/sei-db/proto"
 	"github.com/sei-protocol/sei-chain/sei-db/state_db/ss"
-	"github.com/sei-protocol/sei-chain/sei-db/tools/bench"
+	"github.com/sei-protocol/sei-chain/sei-db/tools/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -71,6 +77,114 @@ func DBWrite(inputKVDir string, numVersions int, outputDir string, dbBackend str
 	if err != nil {
 		panic(err)
 	}
-	bench.BenchmarkDBWrite(backend, inputKVDir, numVersions, concurrency, batchSize)
+	benchmarkDBWrite(backend, inputKVDir, numVersions, concurrency, batchSize)
 	_ = backend.Close()
+}
+
+// benchmarkDBWrite measures random write performance of the db
+// Given an input dir containing all the raw kv data, it writes to the db one version after another
+func benchmarkDBWrite(db types.StateStore, inputKVDir string, numVersions int, concurrency int, batchSize int) {
+	startLoad := time.Now()
+	kvData, err := utils.LoadAndShuffleKV(inputKVDir, concurrency)
+	if err != nil {
+		panic(err)
+	}
+	endLoad := time.Now()
+	fmt.Printf("Finishing loading %+v kv pairs into memory %+v\n", len(kvData), endLoad.Sub(startLoad).String())
+
+	// Write each version sequentially
+	totalTime := time.Duration(0)
+	writeCount := 0
+	for v := 1; v < numVersions+1; v++ {
+		// Write shuffled entries to RocksDB concurrently
+		fmt.Printf("On Version %+v\n", v)
+		startTime := time.Now()
+		latencies := writeToDBConcurrently(db, kvData, concurrency, int64(v), batchSize)
+		endTime := time.Now()
+		totalTime += endTime.Sub(startTime)
+		writeCount += len(latencies)
+
+		sort.Slice(latencies, func(i, j int) bool { return latencies[i] < latencies[j] })
+		// Latencies per version
+		fmt.Printf("P50 Latency: %v\n", utils.CalculatePercentile(latencies, 50))
+		fmt.Printf("P75 Latency: %v\n", utils.CalculatePercentile(latencies, 75))
+		fmt.Printf("P99 Latency: %v\n", utils.CalculatePercentile(latencies, 99))
+		fmt.Printf("Total time: %v\n", totalTime)
+		fmt.Printf("Total Successfully Written %d\n", writeCount)
+		runtime.GC()
+	}
+
+	// Log throughput
+	fmt.Printf("Total Successfully Written %d\n", writeCount)
+	fmt.Printf("Total Time taken: %v\n", totalTime)
+	fmt.Printf("Throughput: %f writes/sec\n", float64(writeCount)/totalTime.Seconds())
+	fmt.Printf("Total records written %d\n", writeCount)
+}
+
+// writeToDBConcurrently generates random write load against the db
+// Given kv pairs (randomly shuffled), the version, batch size, it will spin up `concurrency` goroutines
+// each of which is assigned to a portion of the kv data and writes to db in `batchSize` batches.
+// It maintains a `latencies` channel which aggregates all the latencies
+func writeToDBConcurrently(db types.StateStore, allKVs []utils.KeyValuePair, concurrency int, version int64, batchSize int) []time.Duration {
+	allKVsLen := len(allKVs)
+	allLatencies := make([]time.Duration, 0, allKVsLen)
+	latencies := make(chan time.Duration, allKVsLen)
+
+	kvsPerRoutine := allKVsLen / concurrency
+	remainder := allKVsLen % concurrency
+
+	wg := &sync.WaitGroup{}
+
+	for i := 0; i < concurrency; i++ {
+		wg.Add(1)
+		go func(i int) {
+			defer wg.Done()
+			start := i * kvsPerRoutine
+			end := start + kvsPerRoutine
+
+			if i == concurrency-1 {
+				end += remainder
+			}
+
+			for j := start; j < end; j += batchSize {
+				ncs := &proto.NamedChangeSet{}
+				cs := &proto.ChangeSet{}
+				cs.Pairs = []*proto.KVPair{}
+
+				batchEnd := j + batchSize
+				if batchEnd > end {
+					batchEnd = end
+				}
+
+				// Add key-value pairs to the batch up to batchSize
+				for k := j; k < batchEnd; k++ {
+					kv := allKVs[k]
+					// No store key for benchmarks
+					cs.Pairs = append(cs.Pairs, &proto.KVPair{
+						Key:   kv.Key,
+						Value: kv.Value,
+					})
+				}
+				ncs.Changeset = *cs
+				startTime := time.Now()
+				err := db.ApplyChangesetSync(version, []*proto.NamedChangeSet{ncs})
+				latency := time.Since(startTime)
+
+				if err == nil {
+					latencies <- latency
+				} else {
+					panic(err)
+				}
+			}
+		}(i)
+	}
+
+	wg.Wait()
+	close(latencies)
+
+	for l := range latencies {
+		allLatencies = append(allLatencies, l)
+	}
+
+	return allLatencies
 }
